[ObjC] First pass at bringing back helpers.cc
There are still some things not related to naming in naming.h, but they are
current exposed for gRPC to use.
