+++ src/google/protobuf/compiler/objectivec/BUILD.bazel
@@ -44,6 +44,7 @@ cc_library(
         "field.cc",
         "file.cc",
         "generator.cc",
+        "helpers.cc",
         "map_field.cc",
         "message.cc",
         "message_field.cc",
