+++ src/google/protobuf/compiler/objectivec/extension.cc
@@ -33,6 +33,7 @@
 #include <iostream>
 
 #include "absl/strings/str_cat.h"
+#include "google/protobuf/compiler/objectivec/names.h"
 #include "google/protobuf/compiler/objectivec/helpers.h"
 #include "google/protobuf/descriptor.pb.h"
 #include "google/protobuf/io/printer.h"
