+++ src/google/protobuf/compiler/objectivec/file.cc
@@ -38,6 +38,7 @@
 #include "absl/strings/str_cat.h"
 #include "google/protobuf/compiler/objectivec/enum.h"
 #include "google/protobuf/compiler/objectivec/extension.h"
+#include "google/protobuf/compiler/objectivec/names.h"
 #include "google/protobuf/compiler/objectivec/helpers.h"
 #include "google/protobuf/compiler/objectivec/message.h"
 #include "google/protobuf/io/printer.h"
