+++ src/google/protobuf/compiler/objectivec/map_field.cc
@@ -32,6 +32,7 @@
 #include <string>
 
 #include "google/protobuf/compiler/objectivec/map_field.h"
+#include "google/protobuf/compiler/objectivec/names.h"
 #include "google/protobuf/compiler/objectivec/helpers.h"
 #include "google/protobuf/io/printer.h"
