+++ src/google/protobuf/compiler/objectivec/names.h
@@ -37,7 +37,6 @@
 #include <vector>
 
 #include "google/protobuf/descriptor.h"
-#include "google/protobuf/descriptor.pb.h"
 #include "google/protobuf/io/zero_copy_stream.h"
 
 // clang-format off
@@ -69,19 +68,6 @@ void PROTOC_EXPORT SetProtoPackagePrefixExceptionList(
 // This is only used when UseProtoPackageAsDefaultPrefix() is True.
 std::string PROTOC_EXPORT GetForcedPackagePrefix();
 void PROTOC_EXPORT SetForcedPackagePrefix(const std::string& prefix);
-
-// Generator Prefix Validation Options (see generator.cc for a
-// description of each):
-struct Options {
-  Options();
-  std::string expected_prefixes_path;
-  std::vector<std::string> expected_prefixes_suppressions;
-  bool prefixes_must_be_registered;
-  bool require_prefixes;
-};
-
-// Escape C++ trigraphs by escaping question marks to "\?".
-std::string PROTOC_EXPORT EscapeTrigraphs(absl::string_view to_escape);
 
 // Remove white space from either end of a absl::string_view.
 void PROTOC_EXPORT TrimWhitespace(absl::string_view* input);
@@ -144,107 +130,9 @@ std::string PROTOC_EXPORT OneofEnumName(const OneofDescriptor* descriptor);
 std::string PROTOC_EXPORT OneofName(const OneofDescriptor* descriptor);
 std::string PROTOC_EXPORT OneofNameCapitalized(const OneofDescriptor* descriptor);
 
-// Returns a symbol that can be used in C code to refer to an Objective C
-// class without initializing the class.
-std::string PROTOC_EXPORT ObjCClass(const std::string& class_name);
-
-// Declares an Objective C class without initializing the class so that it can
-// be refrerred to by ObjCClass.
-std::string PROTOC_EXPORT ObjCClassDeclaration(const std::string& class_name);
-
-inline bool HasPreservingUnknownEnumSemantics(const FileDescriptor* file) {
-  return file->syntax() == FileDescriptor::SYNTAX_PROTO3;
-}
-
-inline bool IsMapEntryMessage(const Descriptor* descriptor) {
-  return descriptor->options().map_entry();
-}
-
 // Reverse of the above.
 std::string PROTOC_EXPORT UnCamelCaseFieldName(const std::string& name,
                                                const FieldDescriptor* field);
-
-enum ObjectiveCType {
-  OBJECTIVECTYPE_INT32,
-  OBJECTIVECTYPE_UINT32,
-  OBJECTIVECTYPE_INT64,
-  OBJECTIVECTYPE_UINT64,
-  OBJECTIVECTYPE_FLOAT,
-  OBJECTIVECTYPE_DOUBLE,
-  OBJECTIVECTYPE_BOOLEAN,
-  OBJECTIVECTYPE_STRING,
-  OBJECTIVECTYPE_DATA,
-  OBJECTIVECTYPE_ENUM,
-  OBJECTIVECTYPE_MESSAGE
-};
-
-enum FlagType {
-  FLAGTYPE_DESCRIPTOR_INITIALIZATION,
-  FLAGTYPE_EXTENSION,
-  FLAGTYPE_FIELD
-};
-
-template <class TDescriptor>
-std::string GetOptionalDeprecatedAttribute(const TDescriptor* descriptor,
-                                           const FileDescriptor* file = NULL,
-                                           bool preSpace = true,
-                                           bool postNewline = false) {
-  bool isDeprecated = descriptor->options().deprecated();
-  // The file is only passed when checking Messages & Enums, so those types
-  // get tagged. At the moment, it doesn't seem to make sense to tag every
-  // field or enum value with when the file is deprecated.
-  bool isFileLevelDeprecation = false;
-  if (!isDeprecated && file) {
-    isFileLevelDeprecation = file->options().deprecated();
-    isDeprecated = isFileLevelDeprecation;
-  }
-  if (isDeprecated) {
-    std::string message;
-    const FileDescriptor* sourceFile = descriptor->file();
-    if (isFileLevelDeprecation) {
-      message = sourceFile->name() + " is deprecated.";
-    } else {
-      message = descriptor->full_name() + " is deprecated (see " +
-                sourceFile->name() + ").";
-    }
-
-    std::string result = std::string("GPB_DEPRECATED_MSG(\"") + message + "\")";
-    if (preSpace) {
-      result.insert(0, " ");
-    }
-    if (postNewline) {
-      result.append("\n");
-    }
-    return result;
-  } else {
-    return "";
-  }
-}
-
-std::string PROTOC_EXPORT GetCapitalizedType(const FieldDescriptor* field);
-
-ObjectiveCType PROTOC_EXPORT
-GetObjectiveCType(FieldDescriptor::Type field_type);
-
-inline ObjectiveCType GetObjectiveCType(const FieldDescriptor* field) {
-  return GetObjectiveCType(field->type());
-}
-
-bool PROTOC_EXPORT IsPrimitiveType(const FieldDescriptor* field);
-bool PROTOC_EXPORT IsReferenceType(const FieldDescriptor* field);
-
-std::string PROTOC_EXPORT
-GPBGenericValueFieldName(const FieldDescriptor* field);
-std::string PROTOC_EXPORT DefaultValue(const FieldDescriptor* field);
-bool PROTOC_EXPORT HasNonZeroDefaultValue(const FieldDescriptor* field);
-
-std::string PROTOC_EXPORT
-BuildFlagsString(const FlagType type, const std::vector<std::string>& strings);
-
-// Builds HeaderDoc/appledoc style comments out of the comments in the .proto
-// file.
-std::string PROTOC_EXPORT BuildCommentsString(const SourceLocation& location,
-                                              bool prefer_single_line);
 
 // The name the commonly used by the library when built as a framework.
 // This lines up to the name used in the CocoaPod.
@@ -254,9 +142,24 @@ extern PROTOC_EXPORT const char* const ProtobufLibraryFrameworkName;
 std::string PROTOC_EXPORT
 ProtobufFrameworkImportSymbol(const std::string& framework_name);
 
+// ---------------------------------------------------------------------------
+
+// These aren't really "naming" related, but can be useful for something
+// building on top of ObjC Protos to be able to share the knowledge/enforcement.
+
 // Checks if the file is one of the proto's bundled with the library.
 bool PROTOC_EXPORT
 IsProtobufLibraryBundledProtoFile(const FileDescriptor* file);
+
+// Generator Prefix Validation Options (see generator.cc for a
+// description of each):
+struct Options {
+  Options();
+  std::string expected_prefixes_path;
+  std::vector<std::string> expected_prefixes_suppressions;
+  bool prefixes_must_be_registered;
+  bool require_prefixes;
+};
 
 // Checks the prefix for the given files and outputs any warnings as needed. If
 // there are flat out errors, then out_error is filled in with the first error
@@ -268,29 +171,6 @@ bool PROTOC_EXPORT ValidateObjCClassPrefixes(
 // come from the environment variables.
 bool PROTOC_EXPORT ValidateObjCClassPrefixes(
     const std::vector<const FileDescriptor*>& files, std::string* out_error);
-
-// Generate decode data needed for ObjC's GPBDecodeTextFormatName() to transform
-// the input into the expected output.
-class PROTOC_EXPORT TextFormatDecodeData {
- public:
-  TextFormatDecodeData();
-  ~TextFormatDecodeData();
-
-  TextFormatDecodeData(const TextFormatDecodeData&) = delete;
-  TextFormatDecodeData& operator=(const TextFormatDecodeData&) = delete;
-
-  void AddString(int32_t key, const std::string& input_for_decode,
-                 const std::string& desired_output);
-  size_t num_entries() const { return entries_.size(); }
-  std::string Data() const;
-
-  static std::string DecodeDataForString(const std::string& input_for_decode,
-                                         const std::string& desired_output);
-
- private:
-  typedef std::pair<int32_t, std::string> DataEntry;
-  std::vector<DataEntry> entries_;
-};
 
 // Helper for parsing simple files.
 class PROTOC_EXPORT LineConsumer {
@@ -308,50 +188,6 @@ bool PROTOC_EXPORT ParseSimpleStream(io::ZeroCopyInputStream& input_stream,
                                      const std::string& stream_name,
                                      LineConsumer* line_consumer,
                                      std::string* out_error);
-
-// Helper class for parsing framework import mappings and generating
-// import statements.
-class PROTOC_EXPORT ImportWriter {
- public:
-  ImportWriter(const std::string& generate_for_named_framework,
-               const std::string& named_framework_to_proto_path_mappings_path,
-               const std::string& runtime_import_prefix,
-               bool include_wkt_imports);
-  ~ImportWriter();
-
-  void AddFile(const FileDescriptor* file, const std::string& header_extension);
-  void Print(io::Printer* printer) const;
-
-  static void PrintRuntimeImports(io::Printer* printer,
-                                  const std::vector<std::string>& header_to_import,
-                                  const std::string& runtime_import_prefix,
-                                  bool default_cpp_symbol = false);
-
- private:
-  class ProtoFrameworkCollector : public LineConsumer {
-   public:
-    ProtoFrameworkCollector(std::map<std::string, std::string>* inout_proto_file_to_framework_name)
-        : map_(inout_proto_file_to_framework_name) {}
-
-    virtual bool ConsumeLine(const absl::string_view& line, std::string* out_error) override;
-
-   private:
-    std::map<std::string, std::string>* map_;
-  };
-
-  void ParseFrameworkMappings();
-
-  const std::string generate_for_named_framework_;
-  const std::string named_framework_to_proto_path_mappings_path_;
-  const std::string runtime_import_prefix_;
-  const bool include_wkt_imports_;
-  std::map<std::string, std::string> proto_file_to_framework_name_;
-  bool need_to_parse_mapping_file_;
-
-  std::vector<std::string> protobuf_imports_;
-  std::vector<std::string> other_framework_imports_;
-  std::vector<std::string> other_imports_;
-};
 
 }  // namespace objectivec
 }  // namespace compiler
